Fix directory watch leak
