+++ lib/src/ui/widget/category_pane_item.dart
@@ -12,7 +12,7 @@ import '../../di/app_state/use_paimon.dart';
 import '../../di/fs_watcher.dart';
 import '../util/display_infobar.dart';
 import 'category_drop_target.dart';
-import 'custom_image.dart';
+import 'time_aware_image.dart';
 
 class FolderPaneItem extends PaneItem {
   FolderPaneItem({
